sqlbase: extract partition helper into its own file
structured.go is rather unwieldy; give TranslateValueEncodingToSpan a
proper home.

Release note: None
